[hist] Cast benchmark ranges
This avoids sign conversion warnings.

+++ hist/histv7/benchmark/hist_benchmark_regular.cxx
@@ -2,6 +2,7 @@
 
 #include <benchmark/benchmark.h>
 
+#include <cstddef>
 #include <random>
 #include <vector>
 
@@ -17,7 +18,7 @@ struct RRegularAxis : public benchmark::Fixture {
    {
       std::mt19937 gen;
       std::uniform_real_distribution<> dis;
-      fNumbers.resize(state.range(0));
+      fNumbers.resize(static_cast<std::size_t>(state.range(0)));
       for (std::size_t i = 0; i < fNumbers.size(); i++) {
          fNumbers[i] = dis(gen);
       }
